// backend/src/domain/smart-routing/enums/Season.ts
/**
 * Сезоны для определения доступности транспорта
 * 
 * Используется для валидации сезонности маршрутов:
 * - Летний период: паромы, регулярные рейсы
 * - Зимний период: зимники, ограниченное судоходство
 * - Переходный период: СВП, ледостав/ледоход
 */
export enum Season {
  /**
   * Летний период (1 июня - 18 октября)
   * Доступны: паромы, регулярные рейсы
   */
  SUMMER = 'summer',

  /**
   * Зимний период (1 декабря - 15 апреля)
   * Доступны: зимники, ледовая переправа
   */
  WINTER = 'winter',

  /**
   * Переходный период (16 апреля - 31 мая, 19 октября - 30 ноября)
   * Доступны: СВП, ограниченное судоходство
   */
  TRANSITION = 'transition',

  /**
   * Круглогодично
   * Доступны: авиа, ЖД, автобусы (круглогодичные дороги)
   */
  ALL = 'all',
}